pyproject.toml
[project]
name = "ritrova"
version = "0.2.0"
description = "Ritrova — find again: face and pet recognition for personal photo archives"
requires-python = ">=3.14"
dependencies = [
    "insightface>=0.7.3",
    "onnxruntime>=1.17.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "pillow>=10.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "click>=8.1.0",
    "ultralytics>=8.1.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "sentencepiece>=0.2.1",
    "pi-heif>=1.3.0",
    "python-dotenv>=1.2.2",
    "faiss-cpu>=1.13.2",
    "mlx-vlm>=0.4.0",
    "sacremoses>=0.1.1",
]

[project.scripts]
ritrova = "ritrova.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-playwright>=0.7.2",
    "radon>=6.0.1",
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RET", "T20", "LOG"]
ignore = ["E501", "UP047"]

[tool.ruff.lint.per-file-ignores]
"src/ritrova/app.py" = ["B008"]
"src/ritrova/cli.py" = ["T20"]
"src/ritrova/scanner.py" = ["T20"]
"src/ritrova/scan_pipeline.py" = ["T20"]

[tool.mypy]
python_version = "3.14"
strict = true
ignore_missing_imports = true


[tool.pytest.ini_options]
testpaths = ["tests"]
